+++ Sources/SwiftHTML/Loading/ClientBundlePolicyProviding.swift
@@ -1,0 +1,9 @@
+public protocol ClientBundlePolicyProviding {
+    var clientBundlePolicy: BundlePolicy { get }
+}
+
+public extension ClientBundlePolicyProviding {
+    var clientBundlePolicy: BundlePolicy {
+        .main
+    }
+}